backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.1.0
lxml==6.0.2
openai==1.3.0
numpy
python-dotenv
python-multipart
